Move formatting to respective cleaning scripts
Now that interim data is saved in .rds format, formatting is done
in the relevant cleaning script, which is easier to keep track of.

Only the formatting that must be done after joining the files is now
done after joining.

This also makes the formatting script redundant, so it was removed.